requirements.txt
streamlit
torch
transformers
huggingface_hub>=0.14.1
requests
Pillow